Internal libgeotiff: resync with upstream to fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31636 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/libgeotiff/geo_set.c
@@ -127,7 +127,7 @@ int GTIFKeySet(GTIF *gtif, geokey_t keyID, tagtype_t type, int count,...)
       case TYPE_DOUBLE: dval=va_arg(ap, dblparam_t); val=(char *)&dval;  break;
       case TYPE_ASCII: 
         val=va_arg(ap, char*);
-        count = strlen(val) + 1; /* force = string length */
+        count = (int)strlen(val) + 1; /* force = string length */
         break;
       default:
         assert( FALSE );
